refactor: statistic 원상복구
