+++ src/main/java/com/woowacamp/soolsool/core/statistics/domain/vo/Click.java
@@ -4,21 +4,14 @@ import com.woowacamp.soolsool.core.member.code.MemberErrorCode;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import java.math.BigInteger;
 import java.util.Objects;
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @EqualsAndHashCode
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Click {
 
-    @Column(name = "click", columnDefinition = "decimal(19,2) default 0.0")
-    private BigInteger count;
+    private final BigInteger count;
 
     public Click(final BigInteger count) {
         validateIsValidSize(count);
