--- src/test/resources/statistics.sql
@@ -1,3 +1,0 @@
-INSERT INTO statistics (`statistic_day`, `liquor_id`, `statistic_month`, `statistic_week`, `statistic_year`, `brew_type`, `click`, `impression`, `region`, `sale_price`, `sale_quantity`)
-VALUES
-  (1, 1, 1, 1, 2023, 'Type A', 10.50, 100.00, 'Region X', 20.00, 5.00);
